get rid of cmake stuff

+++ scripts/capnp_build_qnx.sh
@@ -12,6 +12,11 @@ if [[ -z "${LDK_QNX_INSTALL_FOLDER}" ]]; then
     exit 1
 fi
 
+if [[ $# -eq 3 ]]; then
+    echo "You must specify a build type (Release, Debug) and a toolchain file!"
+    exit 1
+fi
+
 . "${QNX_SOURCE_SCRIPT}"
 
 rm -rf build
@@ -21,8 +26,8 @@ cd build
 cmake -G "Unix Makefiles" ..                                                             \
  -DCMAKE_PREFIX_PATH:PATH=${LDK_QNX_INSTALL_FOLDER}                                      \
  -DCMAKE_INSTALL_PREFIX:PATH=${LDK_QNX_INSTALL_FOLDER}                                   \
- -DCMAKE_BUILD_TYPE=Release                                                              \
- -DCMAKE_TOOLCHAIN_FILE=cmake/toolchains/aarch64/qnx/gcc/toolchain_aarch64_qnx_gcc.cmake \
+ -DCMAKE_BUILD_TYPE="$1"                                                                 \
+ -DCMAKE_TOOLCHAIN_FILE="$2"                                                             \
  -DBUILD_TESTING=OFF # off until I figure out what's up with capnp_tool
 
 cmake --build . --target install -- -j "${LUMPDK_NPROC:-$(($(nproc) + 2))}"
